Fix 32-bit and clippy CI failures in NTT module
Use native Word/Lane types throughout pack.rs (instead of u64/u32) so the
same source compiles cleanly on both 32-bit and 64-bit targets. Resolve
the remaining clippy warnings (unnecessary_cast, let_and_return,
too_many_arguments, needless_range_loop, type_complexity) that were
failing the -D warnings CI run.

+++ integer/src/mul/ntt/crt.rs
@@ -216,6 +216,6 @@ mod tests {
         let x = garner_combine::<TripleWord>(&residues[..1], &CRT_INV_IJ, &primes);
         let mut buf = [crate::arch::word::Word::default(); 6];
         x.write_words(&mut buf);
-        assert_eq!(buf[0] as u64, residues[0] as u64);
+        assert_eq!(buf[0], residues[0]);
     }
 }
